LogInViewController rearranged
